worked on graphl

+++ src/graphql/mutations/commentMutation.ts
@@ -2,6 +2,9 @@ import { GraphQLString, GraphQLInt } from "graphql";
 import { CommentType } from "../types/commentType";
 import { ICommentService } from "../../services/abstract/ICommentService";
 import { Container } from "../../provider/repository-service-provider";
+import { currentUser } from "../../http/middlewares/current-user";
+import { requireAuth } from "../../http/middlewares/require-auth";
+import { checkPermission } from "../../http/middlewares/check-permission";
 
 const getCommentService = (): ICommentService => Container.resolve<ICommentService>("CommentService");
 
@@ -12,7 +15,11 @@ export const createComment = {
     rating: { type: GraphQLInt },
     postId: { type: GraphQLInt },
   },
-  resolve: async (_: unknown, args: { description: string; rating: number; postId: number }) => {
+  resolve: async (_: unknown, args: { description: string; rating: number; postId: number }, context:any) => {
+    await currentUser(context.req, context.res, () => {});
+    await requireAuth(context.req, context.res, () => {});
+    await checkPermission('comment.create');
+
     return await getCommentService().create(args);
   },
 };
@@ -25,7 +32,11 @@ export const updateComment = {
     rating: { type: GraphQLInt },
     postId: { type: GraphQLInt },
   },
-  resolve: async (_: unknown, args: {id:number, description: string; rating: number; postId: number }) => {
+  resolve: async (_: unknown, args: {id:number, description: string; rating: number; postId: number }, context:any) => {
+    await currentUser(context.req, context.res, () => {});
+    await requireAuth(context.req, context.res, () => {});
+    await checkPermission('comment.update');
+
     return await getCommentService().update(args.id, args);
   },
 };
@@ -33,7 +44,11 @@ export const updateComment = {
 export const deleteComment = {
   type: GraphQLString,
   args: { id: { type: GraphQLInt } },
-  resolve: async (_: unknown, args: { id: number }) => {
+  resolve: async (_: unknown, args: { id: number }, context:any) => {
+    await currentUser(context.req, context.res, () => {});
+    await requireAuth(context.req, context.res, () => {});
+    await checkPermission('comment.delete');
+
     await getCommentService().delete(args.id);
     return "Comment deleted successfully";
   },
